% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.r
\name{buscar_muni}
\alias{buscar_muni}
\title{Buscar mandados a partir de um municipio}
\usage{
buscar_muni(id_muni, captcha_code = captcha_code_get(), orgao = NULL)
}
\arguments{
\item{id_muni}{codigo do municipio}

\item{captcha_code}{codigo do reCaptcha}

\item{orgao}{codigo do orgao expedidor. Por padrao, busca todos os
orgaos. E necessario somente em municipios que retornam mais de
10.000 resultados. Nesse caso, e necessario buscar os orgaos
individualmente. Para isso, utilize a funcao listar_orgaos().}
}
\value{
tibble com os mandados de prisao
}
\description{
Busca todos os mandados de prisao de um municipio.
}
