% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.r
\name{captcha_code_get}
\alias{captcha_code_get}
\title{Gerar codigo de captcha}
\usage{
captcha_code_get()
}
\value{
codigo do reCaptcha
}
\description{
Gera um codigo do reCaptcha utilizado em outras requisicoes
posteriores. Talvez seja necessario gerar um novo codigo
caso o anterior tenha expirado.
}
