% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.r
\name{listar_municipios}
\alias{listar_municipios}
\title{Listar municipios de um estado}
\usage{
listar_municipios(id_uf, captcha_code = captcha_code_get())
}
\arguments{
\item{id_uf}{codigo do estado}

\item{captcha_code}{codigo do reCaptcha}
}
\value{
tibble com os municipios
}
\description{
Lista todos os municipios de um estado
}
