% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.r
\name{listar_orgaos}
\alias{listar_orgaos}
\title{Listar orgaos expedidores de um municipio}
\usage{
listar_orgaos(id_muni, captcha_code = captcha_code_get())
}
\arguments{
\item{id_muni}{codigo do municipio}

\item{captcha_code}{codigo do reCaptcha}
}
\value{
tibble com os orgaos expedidores
}
\description{
Lista todos os orgaos expedidores de um municipio
}
